+++ api/web-bff/src/main/java/com/zeroskill/web_bff/WebBffApplication.java
@@ -2,8 +2,10 @@ package com.zeroskill.web_bff;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
+@EnableFeignClients(basePackages = "com.zeroskill.web_bff.client")
 public class WebBffApplication {
 
     public static void main(String[] args) {
