+++ api/web-bff/src/main/java/com/zeroskill/web_bff/dto/response/MainPageResponseDto.java
@@ -1,0 +1,8 @@
+package com.zeroskill.web_bff.dto.response;
+
+import com.zeroskill.common.entity.Member;
+
+public record MainPageResponseDto(
+        Member member
+) {
+}
